updated scripts to use .env
