+++ flatfiles/notes.txt
@@ -1,5 +1,4 @@
 //truffle-flattener contracts/ReceiverStorage.sol > flatfiles/flat_ReceiverStorage.sol
-//truffle-flattener contracts/UsingTellor.sol > flatfiles/flat_UsingTellor.sol
-//truffle-flattener contracts/Sender.sol > flatfiles/flat_Sender.sol
-//truffle-flattener contracts/MockTellor.sol > flatfiles/flat_MockTellor.sol
-//truffle-flattener contracts/CentralizedOracle.sol > flatfiles/flat_CentralizedOracle.sol
+//truffle-flattener contracts/Sender.sol > flatfiles/flat_TellorSender.sol
+//truffle-flattener contracts/MockTellor.sol > flatfiles/flat_TellorPlayground.sol
+//truffle-flattener contracts/CentralizedOracle.sol > flatfiles/flat_TellorToo.sol
